Change the heat source handling
The heat source is now a class with subclasses (one for each type of heat sources) that is passed to the Energy class.
This implementation goes along with the new implementation in pTatin3d GENE model.
Add a class for heat source expression when defined on material points.
Update examples.