Add documentation and new class to sum velocities
New class: VelocityCompose(Velocity) inheriting from Velocity designed to create a new velocity field and compute its gradient from the composition of several velocity functions.
The available method to compose functions is, for now, only "sum" which means that the velocity functions will be summed with each others.
Takes a list of Velocity class instances as input for the velocity functions to compose from.

Move some methods in the parent Velocity class.

Add documentation for VelocityLinearAsymmetric and VelocityCompose classes.