Gaussian refactor
Refactoring of the Gaussian class.
Now, the Gaussian class is the parent class of all other Gaussian subclasses.

GaussianConstructor is a class allowing to construct any type of gaussian like distribution. It takes expression(s) and coefficient(s) to construct gaussians in multiple dimensions.

Gaussian2D is a class to construct simple 2D gaussian distributions.

GaussianPlane is a classe to construct gaussian distribution around a given plane

GaussianOptions is a class to gather all the different gaussians created and generate the options.

The former Gaussian() class constructor cannot be used anymore it has to be replaced either by one of the Gaussian2D, GaussianPlane() or GaussianConstructor().
Then all the Gaussian() class instances must be gathered in a list and passed to GaussianOptions() class.